requirements.txt
pyjwt
jupyter
tableauserverclient
vizql-data-service-py
python-dotenv